RulesEngineService and RulesFactory - Refactor rulesfactor and move it to an interface - Create a date filter to use in dataQueryRules
